repos/crypto_zksnark_prover/src/crypto_zksnark.py

import ezkl
import json

class CryptoProver:
    def __init__(self, address, private_key):
        self.address = address
        self.private_key = private_key

    def generate_proof(self):
        raise NotImplementedError("Subclasses must implement this method")

class EthereumProver(CryptoProver):
    def generate_proof(self):
        # Define the circuit for Ethereum key verification
        circuit = {
            "inputs": [
                {"name": "private_key", "type": "field"},
                {"name": "public_address", "type": "field"}
            ],
            "outputs": [{"name": "is_valid", "type": "field"}],
            "circuit": [
                "def main(private_key, public_address):",
                "    computed_address = eth_address(private_key)",
                "    is_valid = computed_address == public_address",
                "    return is_valid"
            ]
        }

        # Compile the circuit
        compiled_circuit = ezkl.compile_circuit(json.dumps(circuit))

        # Generate a proof
        witness = {
            "private_key": self.private_key,
            "public_address": self.address
        }
        proof = ezkl.generate_proof(compiled_circuit, witness)

        return proof

class PolkadotProver(CryptoProver):
    def generate_proof(self):
        # Similar implementation to EthereumProver, but with Polkadot-specific address derivation
        circuit = {
            "inputs": [
                {"name": "private_key", "type": "field"},
                {"name": "public_address", "type": "field"}
            ],
            "outputs": [{"name": "is_valid", "type": "field"}],
            "circuit": [
                "def main(private_key, public_address):",
                "    computed_address = polkadot_address(private_key)",
                "    is_valid = computed_address == public_address",
                "    return is_valid"
            ]
        }

        compiled_circuit = ezkl.compile_circuit(json.dumps(circuit))
        witness = {
            "private_key": self.private_key,
            "public_address": self.address
        }
        proof = ezkl.generate_proof(compiled_circuit, witness)

        return proof

class SolanaProver(CryptoProver):
    def generate_proof(self):
        # Similar implementation to EthereumProver, but with Solana-specific address derivation
        circuit = {
            "inputs": [
                {"name": "private_key", "type": "field"},
                {"name": "public_address", "type": "field"}
            ],
            "outputs": [{"name": "is_valid", "type": "field"}],
            "circuit": [
                "def main(private_key, public_address):",
                "    computed_address = solana_address(private_key)",
                "    is_valid = computed_address == public_address",
                "    return is_valid"
            ]
        }

        compiled_circuit = ezkl.compile_circuit(json.dumps(circuit))
        witness = {
            "private_key": self.private_key,
            "public_address": self.address
        }
        proof = ezkl.generate_proof(compiled_circuit, witness)

        return proof

class BitcoinProver(CryptoProver):
    def generate_proof(self):
        # Similar implementation to EthereumProver, but with Bitcoin-specific address derivation
        circuit = {
            "inputs": [
                {"name": "private_key", "type": "field"},
                {"name": "public_address", "type": "field"}
            ],
            "outputs": [{"name": "is_valid", "type": "field"}],
            "circuit": [
                "def main(private_key, public_address):",
                "    computed_address = bitcoin_address(private_key)",
                "    is_valid = computed_address == public_address",
                "    return is_valid"
            ]
        }

        compiled_circuit = ezkl.compile_circuit(json.dumps(circuit))
        witness = {
            "private_key": self.private_key,
            "public_address": self.address
        }
        proof = ezkl.generate_proof(compiled_circuit, witness)

        return proof
